+++ testOpenCV/Calc/LineCalc.h
@@ -22,7 +22,7 @@ public:
 
 	bool computeIntersectionPoints(cv::Vec4f horizontalLine, const cv::Vec4f& leftBorder, const cv::Vec4f& rightBorder, cv::Vec4f& output) const;
 
-	static void adjustMarkingRect(cv::Rect2f& markingRect, cv::Vec4f& intersectionPoints, double buffer);
+	static void adjustMarkingRect(cv::Rect2d& markingRect, cv::Vec4f& intersectionPoints, double buffer);
 
 	static void adjustBaseLines(cv::Vec4f& baseLines, cv::Vec4f& intersectionPoints, double buffer);
